requirements.txt
django
pyjwt
requests
passlib
pillow
selenium